+++ plugins/lambda-plugin/src/hooks/info.ts
@@ -57,5 +57,6 @@ export default async (context: InfoContext) => {
             }
         }
     }
+
 };